Remove DiffRVRendererAdapter and move update to RVRendererAdapter then fix the sample project
